requirements.txt
# requirements.txt
pandas
pandas-ta
numpy
python-binance
setuptools
google-generativeai
flask
